More robust handling of the camera calibration graph (SPD errors, re-initialization, etc)
CameraCalibrator
- iterate until convergence
- save calibration (.config) and images to /tmp/ in same function call
- don't call draw() internally after adding images

EasyCal and SingleCameraCalibrator
- more robust reinitialization code when adding images (catch SPD errors, iterate until convergence)
- no manual graph iterating or graph iteration threads
- don't reset the camera on re-initialization
- draw only after adding images

EasyCal
- force EasyCal to capture an image

TagMosaic
- make getPositionMeters(column, row) take doubles for column and row so you can add/subtract 0.5 to get the tag border instead of the tag center

+++ java/src/april/tag/TagMosaic.java
@@ -92,7 +92,7 @@ public class TagMosaic
 
     /** Get the position on the tag mosaic (in meters) from the column and row.
       */
-    public double[] getPositionMeters(int col, int row)
+    public double[] getPositionMeters(double col, double row)
     {
         return new double[] { col*tagSpacingMeters ,
                               row*tagSpacingMeters ,
